Add functional test for blocking in Recover
There was a bug, when a task had a job both in ATLAS and Recover and the task
blocked. Both RQs would decrement their nr_running if it was the last (or only)
task; but only one task would see the wakeup.
This has been fixed, but keep the test in case of regressions.